+++ projects/ccd-case-ui-toolkit/src/lib/shared/domain/case-flag/language-interpreter-list-of-values-type.model.ts
@@ -1,0 +1,5 @@
+export interface LanguageInterpreterListOfValuesType {
+  key: string,
+  value: string,
+  value_cy: string
+}