+++ projects/ccd-case-ui-toolkit/src/lib/shared/domain/index.ts
@@ -4,6 +4,7 @@ export * from './alert';
 export * from './case-details.model';
 export * from './case-edit';
 export * from './case-event-data.model';
+export * from './case-flag';
 export * from './case-view';
 export * from './definition';
 export * from './document';